HW5/aggregation-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
clickhouse-driver==0.2.7
psycopg2-binary==2.9.9
apscheduler==3.10.4
boto3==1.34.84
